Implements Behaves
This commit adds implementation detail for `Behaves`, it has two methods
defined:

- `implements` which is meant to be called on the behavior-defining class.
- `behaves_like` which is meant to be called on the behavior-adhering
class.

Take for example relationship between a `Dog` and an `Animal`.

In this case, `Animal` would be the one calling `implements`
(`behavior_defining`) with a list of methods to be implemented,
whereas `Dog` would be calling `behaves_like` to check that it
implements the proper methods.